# src/Ymir.Api/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Ymir.Api | v1",
    "version": "1.0.0"
  },
  "servers": null,
  "paths": {
...
    "/api/conversations/{conversationId}/sites": {
      "post": {
        "tags": [
          "Sites"
        ],
        "operationId": "PublishSite",
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishSiteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sites/{siteId}/publish": {
      "post": {
        "tags": [
          "Sites"
        ],
        "operationId": "RepublishSite",
        "parameters": [
          {
            "name": "siteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepublishSiteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sites/{siteId}/unpublish": {
      "post": {
        "tags": [
          "Sites"
        ],
        "operationId": "UnpublishSite",
        "parameters": [
          {
            "name": "siteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sites/{siteId}": {
      "delete": {
        "tags": [
          "Sites"
        ],
        "operationId": "DeleteSite",
        "parameters": [
          {
            "name": "siteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PublishSiteRequest": {
        "required": [
          "name",
          "sourcePath"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "sourcePath": {
            "type": [
              "null",
              "string"
            ]
          },
          "spaMode": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "RepublishSiteRequest": {
        "required": [
          "conversationId",
          "sourcePath",
          "spaMode"
        ],
        "type": "object",
        "properties": {
          "conversationId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "sourcePath": {
            "type": [
              "null",
              "string"
            ]
          },
          "spaMode": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
      "SiteAccessMode": {
        "enum": [
          "Public",
          "AllUsers",
          "SelectedUsers"
        ]
      },
      "SiteResponse": {
        "required": [
          "id",
          "name",
          "url",
          "status",
          "accessMode",
          "spaMode",
          "conversationId",
          "sourcePath",
          "fileCount",
          "totalBytes",
          "publishedAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": [
              "null",
              "string"
            ],
            "format": "uri"
          },
          "status": {
            "$ref": "#/components/schemas/SiteStatus"
          },
          "accessMode": {
            "$ref": "#/components/schemas/SiteAccessMode"
          },
          "spaMode": {
            "type": "boolean"
          },
          "conversationId": {
            "type": "string",
            "format": "uuid"
          },
          "sourcePath": {
            "type": "string"
          },
          "fileCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "publishedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SiteStatus": {
        "enum": [
          "Unpublished",
          "Published"
        ]
      },
...
    }
  },
...
}